refactor: move getTransactions service to a new location
